EX 3

+++ src/Ex3/Video.java
@@ -2,7 +2,7 @@ package Ex3;
 
 public class Video extends Ouvrage{
     private String editeur;
-    private int durée;
+    private float durée;
     public void afficher(){
         super.afficher();
         System.out.println("Editeur : "+editeur+"\n"+"Durée : "+durée);
